ov016: name vtable slots for RockPillar (0, 6) and FloatOnWaterPlatformJrb (0)
Verified from config/arm9/overlays/ov016/relocs.txt: RockPillar's vtable
(_ZTV10RockPillar) relocates slot 0 to 0x02112e1c and slot 6 to 0x02112b50;
FloatOnWaterPlatformJrb's (via its base daObjFloatBoard_c, slot 0 only --
the base itself supplies slot 3) relocates slot 0 to 0x02112fa8.
dBgActor_c's own vtable carries different words at every offset --
genuine overrides.

func_ov016_02112e1c -> _ZN10RockPillar13InitResourcesEv        (slot 0)
func_ov016_02112b50 -> _ZN10RockPillar8BehaviorEv               (slot 6)
func_ov016_02112fa8 -> _ZN23FloatOnWaterPlatformJrb13InitResourcesEv (slot 0)

Both RockPillar functions are minimal-diff member-function migrations: the
pre-migration `char *c` self pointer becomes a local `char *c = (char *)
this;`, everything else byte-for-byte unchanged, including Behavior's
`#pragma opt_common_subs off` / `inline_fn` address-preserving idiom for
an 8-slot state machine (actor pointer, state byte, counter byte, two s16
angle/velocity fields at 0x31e-0x327) that RockPillar.h declares only as
opaque `pad_320[0x8]` -- naming those fields is out of this task's scope,
so the raw-offset recovery is kept verbatim rather than guessed at.
InitResources still goes through the shadow struct daObjKi_Hasira_c
(include/daObjKi_Hasira_c.h, the RTTI class name) for the two named
fields it does touch (unk_08e = dActor_c::mAngleY, unk_0cc =
dActor_c::mAreaId). FloatOnWaterPlatformJrb::InitResources is a
cross-overlay tail-call veneer to daObjFloatBoard_c's shared,
still-unmigrated file-load helper func_ov002_020b5e58 -- out of scope,
kept under its existing name (mwccarm-codegen.md's long_calls veneer
idiom).

TRAP HIT AND FIXED: RockPillar::InitResources' first draft left its extern
declarations un-namespaced in the //cpp file, which double-mangled every
one of them (double-mangling-defect) -- build_pin.verify still passed
(the compiled call-site bytes are identical either way, mangled name or
not) but eligible.py's isolation step then reported "unresolvable:
_Z19func_ov016_02112a9cPc,..." for a function that never showed up as a
compile error. Wrapped all the free-function externs in one `extern "C"
{ }` block; re-verified byte-exact and eligible afterward.

Gates: build_pin.verify True/2004b56 for all 3, re-checked after restoring
through the manual checkout-swap bracket (no git stash). eligible.py
name-list diff is exactly these 3 renames, nothing else gained or lost
(11026/11170 both sides, after the extern-C fix above -- the first
attempt at this bracket read 11025/11170, one function short, which is
what surfaced the double-mangling defect).
