+++ config/arm9/overlays/ov016/symbols.txt
@@ -33,12 +33,12 @@ _ZN10RockPillarD0Ev kind:function(arm,size=0x58) addr:0x02112a44
 func_ov016_02112a9c kind:function(arm,size=0x48) addr:0x02112a9c
 _ZN10RockPillar16CleanupResourcesEv kind:function(arm,size=0x44) addr:0x02112ae4
 _ZN10RockPillar6RenderEv kind:function(arm,size=0x28) addr:0x02112b28
-func_ov016_02112b50 kind:function(arm,size=0x2cc) addr:0x02112b50
-func_ov016_02112e1c kind:function(arm,size=0xa8) addr:0x02112e1c
+_ZN10RockPillar8BehaviorEv kind:function(arm,size=0x2cc) addr:0x02112b50
+_ZN10RockPillar13InitResourcesEv kind:function(arm,size=0xa8) addr:0x02112e1c
 RockPillar_Spawn kind:function(arm,size=0x30) addr:0x02112ec4
 _ZN23FloatOnWaterPlatformJrbD1Ev kind:function(arm,size=0x50) addr:0x02112ef4
 _ZN23FloatOnWaterPlatformJrbD0Ev kind:function(arm,size=0x64) addr:0x02112f44
-func_ov016_02112fa8 kind:function(arm,size=0x14) addr:0x02112fa8
+_ZN23FloatOnWaterPlatformJrb13InitResourcesEv kind:function(arm,size=0x14) addr:0x02112fa8
 FloatOnWaterPlatformJrb_Spawn kind:function(arm,size=0x3c) addr:0x02112fbc
 _ZN10SlidingBoxD1Ev kind:function(arm,size=0x4c) addr:0x02112ff8
 _ZN10SlidingBoxD0Ev kind:function(arm,size=0x60) addr:0x02113044
