+++ config/match_attempts.jsonl
@@ -171,9 +171,9 @@
 {"schemaVersion":1,"functionId":"ov014:0x0211236c","id":"ov014:0x0211236c","attemptId":"93851c8b077846409127f1385ad66b24","parentAttemptId":"b5a0ae27cf1e43d49b80e42faf881c49","module":"ov014","addr":34677612,"name":"func_ov014_0211236c","status":"no_progress","kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build","author":"lunavyqo","divergences":2,"prevBestDivergences":2,"improvedNearMiss":false,"srcPath":"scratch/func_ov014_0211236c_div2.c","note":"tail store/sub scheduling grind (volatile/pragmas/fake-deps) stuck at div=2","usedNearMissDraft":true,"usedGhidraDraft":false,"base":{"kind":"previous_attempt","attemptId":"b5a0ae27cf1e43d49b80e42faf881c49","divergences":2},"sessionScope":"focused","batchSize":1}
 {"schemaVersion":1,"functionId":"ov014:0x0211236c","id":"ov014:0x0211236c","attemptId":"abbfaa312e894b8c8b3965ca9181750d","parentAttemptId":"b5a0ae27cf1e43d49b80e42faf881c49","module":"ov014","addr":34677612,"name":"func_ov014_0211236c","status":"matched","kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build","author":"lunavyqo","divergences":0,"prevBestDivergences":2,"improvedNearMiss":true,"srcPath":"src/func_ov014_0211236c.c","note":"MATCH: sum.x/y/z assigns then i=i-1 then cur-=0xc (not i between sum.z); linkcheck VERIFIED","usedNearMissDraft":true,"usedGhidraDraft":false,"base":{"kind":"previous_attempt","attemptId":"b5a0ae27cf1e43d49b80e42faf881c49","divergences":2},"sessionScope":"focused","batchSize":1}
 {"schemaVersion":1,"functionId":"ov016:0x02112b50","id":"ov016:0x02112b50","attemptId":"66682c7523e84c6e93db0bc3630aa2a0","parentAttemptId":null,"module":"ov016","addr":34679632,"name":"func_ov016_02112b50","status":"no_progress","kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build","author":"lunavyqo","divergences":10,"prevBestDivergences":10,"improvedNearMiss":false,"srcPath":"/tmp/match_work/func_ov016_02112b50.c","note":"baseline near-miss draft confirm: case-2 pure regperm p322 r1 vs r3 / base r3 vs r0 (div=10 size-exact)","usedNearMissDraft":true,"usedGhidraDraft":false,"base":{"kind":"near_miss_draft","divergences":10},"sessionScope":"focused","batchSize":1}
-{"schemaVersion":1,"functionId":"ov016:0x02112b50","id":"ov016:0x02112b50","attemptId":"3c931ecbc1954214a79263384c1bc9c3","parentAttemptId":"66682c7523e84c6e93db0bc3630aa2a0","module":"ov016","addr":34679632,"name":"func_ov016_02112b50","status":"matched","kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build","author":"lunavyqo","divergences":0,"prevBestDivergences":10,"improvedNearMiss":true,"srcPath":"src/func_ov016_02112b50.c","note":"MATCH: case-2 compound += with inline ((s16*)(c+0x300))[i] RHS (no named base); ov092-style regperm lever; linkcheck VERIFIED","usedNearMissDraft":true,"usedGhidraDraft":false,"base":{"kind":"previous_attempt","attemptId":"66682c7523e84c6e93db0bc3630aa2a0","divergences":10},"sessionScope":"focused","batchSize":1}
+{"schemaVersion":1,"functionId":"ov016:0x02112b50","id":"ov016:0x02112b50","attemptId":"3c931ecbc1954214a79263384c1bc9c3","parentAttemptId":"66682c7523e84c6e93db0bc3630aa2a0","module":"ov016","addr":34679632,"name":"func_ov016_02112b50","status":"matched","kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build","author":"lunavyqo","divergences":0,"prevBestDivergences":10,"improvedNearMiss":true,"srcPath":"src/_ZN10RockPillar8BehaviorEv.cpp","note":"MATCH: case-2 compound += with inline ((s16*)(c+0x300))[i] RHS (no named base); ov092-style regperm lever; linkcheck VERIFIED","usedNearMissDraft":true,"usedGhidraDraft":false,"base":{"kind":"previous_attempt","attemptId":"66682c7523e84c6e93db0bc3630aa2a0","divergences":10},"sessionScope":"focused","batchSize":1}
 {"schemaVersion":1,"functionId":"ov014:0x0211236c","id":"ov014:0x0211236c","attemptId":"0c1e0cdb6afd44a7b57564768a47f40c","parentAttemptId":null,"module":"ov014","addr":34677612,"name":"func_ov014_0211236c","status":"matched","kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build","improvedNearMiss":false,"srcPath":"src/func_ov014_0211236c.c","base":{"kind":"scratch"},"sessionScope":"batch","batchSize":2}
-{"schemaVersion":1,"functionId":"ov016:0x02112b50","id":"ov016:0x02112b50","attemptId":"575676374bee4e77a3de050e6839bf5a","parentAttemptId":null,"module":"ov016","addr":34679632,"name":"func_ov016_02112b50","status":"matched","kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build","improvedNearMiss":false,"srcPath":"src/func_ov016_02112b50.c","base":{"kind":"scratch"},"sessionScope":"batch","batchSize":2}
+{"schemaVersion":1,"functionId":"ov016:0x02112b50","id":"ov016:0x02112b50","attemptId":"575676374bee4e77a3de050e6839bf5a","parentAttemptId":null,"module":"ov016","addr":34679632,"name":"func_ov016_02112b50","status":"matched","kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build","improvedNearMiss":false,"srcPath":"src/_ZN10RockPillar8BehaviorEv.cpp","base":{"kind":"scratch"},"sessionScope":"batch","batchSize":2}
 {"schemaVersion":1,"functionId":"ov014:0x0211250c","id":"ov014:0x0211250c","attemptId":"0a14a928c6684ab89d95763f1cd7725d","parentAttemptId":null,"module":"ov014","addr":34678028,"name":"func_ov014_0211250c","status":"near_miss","kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build","author":"lunavyqo","divergences":9,"prevBestDivergences":181,"improvedNearMiss":true,"srcPath":"/tmp/func_ov014_0211250c_best.c","note":"control-flow fix (always matrix); regalloc to r8/sl; remaining: prologue seg/cmp schedule + Vec3 copy 3-reg load order","usedNearMissDraft":true,"usedGhidraDraft":false,"base":{"kind":"near_miss_draft","divergences":181},"sessionScope":"focused","batchSize":1}
 {"schemaVersion":1,"functionId":"ov014:0x0211250c","id":"ov014:0x0211250c","attemptId":"b53f8ac577dc48cba3b4c259bc81f3de","parentAttemptId":null,"module":"ov014","addr":34678028,"name":"func_ov014_0211250c","status":"matched","kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build","author":"lunavyqo","improvedNearMiss":false,"srcPath":"src/func_ov014_0211250c.c","note":"banked","base":{"kind":"scratch"},"sessionScope":"focused","batchSize":1}
 {"schemaVersion":1,"functionId":"ov014:0x0211250c","id":"ov014:0x0211250c","attemptId":"a2ab9896623b4c2d957a8c7fa73b31e3","parentAttemptId":"0a14a928c6684ab89d95763f1cd7725d","module":"ov014","addr":34678028,"name":"func_ov014_0211250c","status":"matched","kind":"ai","model":"grok-4.5","reasoning":"high","harness":"grok-build","author":"lunavyqo","improvedNearMiss":false,"srcPath":"src/func_ov014_0211250c.c","note":"PERM_LINESWAP setup order (seg before state) closed final 2-word residual","usedNearMissDraft":true,"usedGhidraDraft":false,"base":{"kind":"previous_attempt","attemptId":"0a14a928c6684ab89d95763f1cd7725d","divergences":9},"sessionScope":"focused","batchSize":1}
