+++ include/RockPillar.h
@@ -21,7 +21,9 @@ struct RockPillar : dBgActor_c {
 
     virtual ~RockPillar();            /* slots 16 (D1), 17 (D0) */
 
+    virtual s32   InitResources();         /* slot  0 */
     virtual s32   CleanupResources();      /* slot  3 */
+    virtual s32   Behavior();              /* slot  6 */
     virtual s32   Render();                /* slot  9 */
 };
