--- src/func_ov016_02112fa8.c
@@ -1,11 +1,0 @@
-// @symbol func_ov016_02112fa8
-// recovered name: daObjKi_Ita_c_InitResources
-/* recovered: renamed to Class_Method, declarations from a shared header */
-#include "decl_common.h"
-/* recovered: renamed to Class_Method */
-/* daObjKi_Ita_c::InitResources - recovered from vtable slot identity */
-
-int func_ov016_02112fa8(void *self)
-{
-    return func_ov002_020b5e58(self, &data_ov016_02114b8c);
-}
